Some changes and removed the cli feature for datasets which was never working or used
