+++ advsecurenet/datasets/dataset_factory.py
@@ -1,4 +1,5 @@
-from typing import Optional, Dict
+from typing import Optional, Dict, Any
+import warnings
 
 from advsecurenet.datasets.base_dataset import BaseDataset
 from advsecurenet.datasets.Cifar10 import CIFAR10Dataset, CIFAR100Dataset
@@ -53,14 +54,19 @@ class DatasetFactory:
                 dataset_provider = DatasetFactory._create_provider(split_config, dataset_type)
 
                 # 2. Prepare kwargs for the `load_dataset` method
-                load_kwargs = split_config.kwargs.copy()
-                load_kwargs['dataset_name'] = identifier
-                load_kwargs['split'] = split_config.source_split_name
-                load_kwargs['root'] = getattr(split_config, 'path', None)
-                load_kwargs['download'] = getattr(resolved_config, 'download', True)
-                
-                # Merge runtime kwargs last
-                load_kwargs.update(runtime_kwargs)
+                keys_to_override = {
+                    'dataset_name': identifier,
+                    'split': split_config.source_split_name,
+                    'root': getattr(split_config, 'path', None),
+                    'download': getattr(resolved_config, 'download', True)
+                }
+
+                load_kwargs = _prepare_load_kwargs(
+                    base_kwargs=split_config.kwargs,
+                    keys_to_override=keys_to_override,
+                    runtime_kwargs=runtime_kwargs,
+                    logical_name=logical_name
+                )
 
                 # 3. Process kwargs and load the dataset
                 processed_load_kwargs = dataset_provider.process_kwargs_load_dataset(load_kwargs)
@@ -152,10 +158,81 @@ def _infer_dataset_class_from_type(dataset_type: DatasetType) -> type:
         return DATASET_MAP[dataset_type]
     except KeyError:
         raise ValueError(f"Unknown dataset type: {dataset_type}")
+    
+def _merge_dicts_with_warning(
+    base_dict: Dict[str, Any],
+    override_dict: Dict[str, Any],
+    warning_template: str
+) -> Dict[str, Any]:
+    """
+    Merges an override dictionary into a copy of a base dictionary,
+    issuing a formatted warning on any conflicts. Returns a new dictionary.
+    
+    Args:
+        base_dict (Dict): The dictionary with default values.
+        override_dict (Dict): The dictionary with new values to merge.
+        warning_template (str): A format-string for the warning message.
+
+    Returns:
+        Dict[str, Any]: A new dictionary containing the merged result.
+    """
+    merged = base_dict.copy()
+    for key, value in override_dict.items():
+        if key in merged:
+            warnings.warn(
+                warning_template.format(
+                    key=key,
+                    old_value=merged.get(key),
+                    new_value=value
+                )
+            )
+        merged[key] = value
+    return merged
+
+def _prepare_load_kwargs(
+    base_kwargs: Dict[str, Any],
+    keys_to_override: Dict[str, Any],
+    runtime_kwargs: Dict[str, Any],
+    logical_name: str
+) -> Dict[str, Any]:
+    """
+    Prepares the final keyword arguments for the load_dataset method by merging
+    different sources of arguments and warning on conflicts.
+    
+    Args:
+        base_kwargs (Dict): The initial kwargs from the configuration.
+        keys_to_override (Dict): Dictionary of keys that must be set (e.g., {'dataset_name': 'cifar10'}).
+        runtime_kwargs (Dict): Additional kwargs passed at runtime.
+        logical_name (str): The logical name of the split (e.g., 'train') for warning messages.
+
+    Returns:
+        Dict[str, Any]: The final, processed dictionary of keyword arguments.
+    """
+    # 1. Start with the base kwargs and override with resolved config keys
+    load_kwargs = _merge_dicts_with_warning(
+        base_dict=base_kwargs,
+        override_dict=keys_to_override,
+        warning_template=(
+            f"The 'dataset_kwargs' for split '{logical_name}' contains a '{{key}}' key. "
+            f"It will be overridden by the resolved value '{{new_value}}'."
+        )
+    )
+
+    # 2. Take the result and override with runtime kwargs from the CLI call
+    final_kwargs = _merge_dicts_with_warning(
+        base_dict=load_kwargs,
+        override_dict=runtime_kwargs,
+        warning_template=(
+            f"Runtime argument '{{key}}' is overriding a configuration value for split '{logical_name}'. "
+            f"Old: '{{old_value}}', New: '{{new_value}}'"
+        )
+    )
+    
+    return final_kwargs
 
 
-    @staticmethod
-    def available_datasets() -> list:
+@staticmethod
+def available_datasets() -> list:
         """
         Returns a list of available datasets.
