--- cli/commands/utils/datasets/__init__.py
@@ -1,1 +1,0 @@
-from .commands import datasets
